Go steady mode after `_delay_input` for most modes
It's kind of irritating to have to either
    a. go through all the remaining, or
    b. cold boot the light
just to get back to steady mode.

It should be pretty uncommon to want to cycle leisurely through blinky
modes as well, so we're not losing much with this change. The main
functional difference is that the led will be off during the input
delays instead of flashing out the first couple loops of the
blinky modes.